refactor: added Home page
