+++ package.json
@@ -32,6 +32,7 @@
 		"react-dnd-html5-backend": "^16.0.1",
 		"react-dom": "^18.2.0",
 		"react-redux": "^9.2.0",
+		"react-router-dom": "^7.6.1",
 		"redux": "^5.0.1"
 	},
 	"devDependencies": {
